+++ tests/e2e/modules/clock_es_spec.js
@@ -78,4 +78,27 @@ describe("Clock set to spanish language module", function () {
 				.getText(".clock .time").should.eventually.match(timeRegex);
 		});
 	});
+
+
+	describe("with showWeek config enabled", function() {
+		before(function() {
+			// Set config sample for use in test
+			process.env.MM_CONFIG_FILE = "tests/configs/modules/clock/es/clock_showWeek.js";
+		});
+
+		beforeEach(function (done) {
+			app.start().then(function() { done(); } );
+		});
+
+		afterEach(function (done) {
+			app.stop().then(function() { done(); });
+		});
+
+		it("shows week with correct format", function() {
+			const weekRegex = /^Semana [0-9]{1,2}$/;
+			return app.client.waitUntilWindowLoaded()
+				.getText(".clock .week").should.eventually.match(weekRegex);
+		});
+	});
+
 });